Adding return to iterate.onItem to increment "recordCount" only if it returns "undefined" (keep compatibility) or "true".
Because IndexedDB is limited on filtering, I had to use iterate and manually check each item with custom filters.
But I still need to limit the quantity of records returned, so I don't iterate the whole store, I know that there is the param "limit" the problem is that IDBWrapper doesn't know how many records passed my filter.
To solve this, I changed the "onItem" function to return a boolean and increment "recordCount" only if it returns "undefined" (keep compatibility) or "true".