Refactor to receive the param "dbName" instead of "storePrefix" and an array of stores.
Creation of a class called "Store" which contains the specific attributes of a store.
Main class changed to receive the param dbName and an array of store params to create each store inside the DB.
With these changes the user is able to create one DB with multiple stores.
Tests changed to comply with the new behavior.